Updated to fix Popover on Ipad size wrong keyboard is hiding
Scenario:
You have a UITextField that you want to populate using a
ActionsheetPicker.
You have other UITextFields too.
Picking up touch of the textfield in textFieldShouldBeginEditing,
resigining first responder (to hide the keyboard) and displaying the
ActionSheetPicker instead.

On Ipad the source uses a Popover instead.

If the keyboard is still showing the popup is sized wrongly,
hiding the picker.

To fix this I added a NSNotification if the keyboarddidhide to
redisplay the picker.
However the popover does not calculate its size properly until
the next run loop so I have to use a performSelector withDelay
before actually redisplaying the popover.

	modified:   ActionSheetPicker.m